tests/hwtest: Fix implicit conversion from float to double
Same as d520b3f.

+++ tests/hwtest/adcs/src/temp_test.c
@@ -35,7 +35,7 @@ static int temp_obc_test(uint32_t *err_cnt)
 			all_ret = -1;
 			continue;
 		}
-		LOG_INF("%s Temperature: %.4f [deg]", obc_pos_name[i], temp);
+		LOG_INF("%s Temperature: %.4f [deg]", obc_pos_name[i], (double)temp);
 	}
 
 	return all_ret;
@@ -51,7 +51,7 @@ static int temp_xadc_test(uint32_t *err_cnt)
 		LOG_ERR("OBC XADC Temperature: Failed");
 		(*err_cnt)++;
 	} else {
-		LOG_INF("OBC XADC Temperature: %.4f [deg]", temp);
+		LOG_INF("OBC XADC Temperature: %.4f [deg]", (double)temp);
 	}
 
 	return ret;
@@ -81,7 +81,7 @@ static int temp_adcs_test(uint32_t *err_cnt)
 			all_ret = -1;
 			continue;
 		}
-		LOG_INF("%s Temperature: %.4f [deg]", adcs_pos_name[i], temp);
+		LOG_INF("%s Temperature: %.4f [deg]", adcs_pos_name[i], (double)temp);
 	}
 
 	return all_ret;
