+++ tests/hwtest/main/src/cv_test.c
@@ -97,7 +97,7 @@ static int cv_xadc_test(uint32_t *err_cnt)
 			all_ret = -1;
 			continue;
 		}
-		LOG_INF("%s: %.4f [v]", xadc_pos_name[i], cv);
+		LOG_INF("%s: %.4f [v]", xadc_pos_name[i], (double)cv);
 	}
 
 	return all_ret;
